+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Reduce gdal warning about projection strings (#559)
 - Don't report needless frame information for some single frame files (#558)
 
+### Changes
+- The PIL tilesource priority is slightly higher than other fallback sources (#560)
+
 ### Bug Fixes
 - Fix compositing when using different frame numbers and partial tiles (#557)